# api-gateway/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  booking.CreatePaymentRequest:
    properties:
      amount:
        type: number
      booking_id:
        type: string
      payment_method:
        type: string
      transaction_id:
        type: string
      user_id:
        type: string
    type: object
...
  booking.ListPaymentsResponse:
    properties:
      payments:
        items:
          $ref: '#/definitions/booking.PaymentResponse'
        type: array
    type: object
...
  booking.NewNotification:
    properties:
      message:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  booking.Notification:
    properties:
      created_at:
        type: string
      id:
        type: string
      message:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  booking.PaymentResponse:
    properties:
      amount:
        type: number
      booking_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      payment_method:
        type: string
      status:
        type: string
      transaction_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Gateway of On-Demand Car Wash Service
  title: On-Demand Car Wash Service
  version: "1.0"
paths:
...
  /admin/payments/{id}:
    get:
      description: Retrieves payment details by payment ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.PaymentResponse'
        "400":
          description: Invalid payment ID
          schema:
            type: string
        "404":
          description: Payment not found
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieves a payment by ID
      tags:
      - payments
...
  /customer/payments:
    get:
      description: Retrieves a list of payments with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Page limit
        in: query
        name: limit
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListPaymentsResponse'
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Lists all payments with pagination
      tags:
      - payments
    post:
      description: Processes a payment and returns the payment details
      parameters:
      - description: Payment details
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/booking.CreatePaymentRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.PaymentResponse'
        "400":
          description: Invalid payment data
          schema:
            type: string
        "500":
          description: Server error while processing payment
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Creates a new payment
      tags:
      - payments
  /customer/payments/{id}:
    get:
      description: Retrieves payment details by payment ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.PaymentResponse'
        "400":
          description: Invalid payment ID
          schema:
            type: string
        "404":
          description: Payment not found
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieves a payment by ID
      tags:
      - payments
...
  /provider/notifications:
    post:
      description: Adds a new notification
      parameters:
      - description: Receiver ID, Title and Message
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/booking.NewNotification'
      responses:
        "201":
          description: Notification created
          schema:
            type: string
        "400":
          description: Invalid data format
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Creates notification
      tags:
      - notification
  /provider/notifications/{id}:
    get:
      description: Gets notification by ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.Notification'
        "400":
          description: Invalid data format
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets notification
      tags:
      - notification
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"